renamed directory from helpers to clients
