+++ examples/credential/ws/main.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/Peersyst/xrpl-go/examples/helpers"
+	"github.com/Peersyst/xrpl-go/examples/clients"
 	"github.com/Peersyst/xrpl-go/pkg/crypto"
 	rippleTime "github.com/Peersyst/xrpl-go/xrpl/time"
 	"github.com/Peersyst/xrpl-go/xrpl/transaction"
@@ -17,7 +17,7 @@ func main() {
 
 	fmt.Println("⏳ Setting up client...")
 
-	client := helpers.GetDevnetWebsocketClient()
+	client := clients.GetDevnetWebsocketClient()
 	fmt.Println("Connecting to server...")
 	if err := client.Connect(); err != nil {
 		fmt.Println(err)
@@ -87,7 +87,7 @@ func main() {
 		URI:            hex.EncodeToString([]byte("https://example.com")),
 	}
 
-	helpers.SubmitAndWait(client, txn, issuer)
+	clients.SubmitAndWait(client, txn, issuer)
 
 	// -----------------------------------------------------
 
@@ -102,7 +102,7 @@ func main() {
 		Issuer:         types.Address(issuer.ClassicAddress),
 	}
 
-	helpers.SubmitAndWait(client, acceptTxn, subjectWallet)
+	clients.SubmitAndWait(client, acceptTxn, subjectWallet)
 
 	// -----------------------------------------------------
 
@@ -118,5 +118,5 @@ func main() {
 		Subject:        types.Address(subjectWallet.ClassicAddress),
 	}
 
-	helpers.SubmitAndWait(client, deleteTxn, issuer)
+	clients.SubmitAndWait(client, deleteTxn, issuer)
 }
